Added an error catch
In the code to generate a random line, I added a try/except block and a dump to a log file so I can see what went wrong.

It also plays an obvious tone when it happens so I'll know to check the log file (randomLineError.txt)

Also adjusted the FULL_URL_WEIGHT in config.py to more heavily weight the FULL_URL entry since it's multiplicative.